+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.5] - 2026-01-12
+
+### Fixed
+- Provisioner ID mismatch causes range create status update to fail
+- Range subnets have no route to s3 for agent downloads
+
 ## [1.0.4] - 2026-01-12
 
 ### Changed